5.1.13 Перенесет рендеринг доски и карточек из main.js в controllers/board.js
