--- src/controllers/board-controller.js
@@ -1,8 +1,0 @@
-export default class BoardController {
-  constructor(container) {
-
-  }
-  render(tasks) {
-
-  }
-}
